adding if to prevent from running on local dev laptops during app setup

+++ recipes/default.rb
@@ -38,19 +38,19 @@ if node['iis']['components']
       all false
       source node['iis']['source'] unless node['iis']['source'].nil?
       install_method :windows_feature_powershell
-      guard_interpreter :powershell_script
-      not_if "(Get-WindowsFeature -Name #{feature}).InstallState -eq \"Installed\""
     end
   end
 end
 
 list_of_features = (node['iis']['components'].map { |feature| "'#{feature}'" }).join(", ")
 
-reboot 'iis install reboot' do
-  action :reboot_now
-  reason 'Required reboot to finish installation of iis windows features'
-  guard_interpreter :powershell_script
-  not_if "((Get-WindowsFeature -Name #{list_of_features}).InstallState -eq \"Installed\").Count -eq #{node['iis']['components'].count}"
+if (node.chef_environment != "_default") 
+  reboot 'iis install reboot' do
+    action :reboot_now
+    reason 'Required reboot to finish installation of iis windows features'
+    guard_interpreter :powershell_script
+    not_if "((Get-WindowsFeature -Name #{list_of_features}).InstallState -eq \"Installed\").Count -eq #{node['iis']['components'].count}"
+    end
 end
 
 service 'iis' do
